rename & design : 페이지 컴포넌트단위로 폴더 수정, eiditor ui 추가
