+++ package.json
@@ -11,6 +11,7 @@
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "@toast-ui/react-editor": "^3.2.3",
     "ace-builds": "^1.32.8",
     "axios": "^1.6.8",
     "body-parse": "^0.1.0",
