--- src/pages/CodeCreate.jsx
@@ -1,8 +1,0 @@
-import { AnnotationCode } from "../components/AnnotationCode/AnnotationCode";
-export const CodeCreate = () => {
-  return (
-    <div>
-      <AnnotationCode />
-    </div>
-  );
-};
