+++ src/pages/CodeCreate/CodeCreate.jsx
@@ -1,0 +1,10 @@
+import { AnnotationCode } from './AnnotationCode';
+import { ToastEditor } from './ToastEditor';
+export const CodeCreate = () => {
+  return (
+    <div>
+      <AnnotationCode />
+      <ToastEditor />
+    </div>
+  );
+};
