+++ src/pages/Main/Main.jsx
@@ -1,4 +1,4 @@
-import { ImageGallery } from "../components/imagegallery/ImageGallery";
+import { ImageGallery } from '../../components/imagegallery/ImageGallery';
 export const Main = () => {
   return (
     <div>
