--- src/pages/PostDetail.jsx
@@ -1,8 +1,0 @@
-import { PostDetailComponent } from "../components/imagegallery/ImageGalleryComponents/PostDetailComponent";
-export const PostDetail = () => {
-  return (
-    <div>
-      <PostDetailComponent />
-    </div>
-  );
-};
