+++ src/pages/PostDetail/PostDetail.jsx
@@ -1,0 +1,8 @@
+import { PostDetailComponent } from '../../components/imagegallery/ImageGalleryComponents/PostDetailComponent';
+export const PostDetail = () => {
+  return (
+    <div>
+      <PostDetailComponent />
+    </div>
+  );
+};
